requirements.txt

streamlit>=1.44.1
openai>=1.76.0
python-dotenv>=1.0.0
